pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pylogix>=1.0.6",
    "python-snap7>=2.0.2",
]
